+++ R/GetRegionOfInterest.R
@@ -3,11 +3,14 @@
 #' Create a spatial polygon describing the convex hull of a set of spatial points.
 #'
 #' @param obj 'SpatialPoints*'.
-#'   Spatial points
+#'   Sample of points
 #' @param alpha 'numeric' number.
-#'   Value of alpha, used to implement a generalization of the convex hull
+#'   Value of \eqn{\alpha}, used to implement a generalization of the convex hull
 #'   (Edelsbrunner and others, 1983).
+#'   As \eqn{\alpha} decreases, the shape shrinks.
 #'   Requires that the \pkg{alphahull} and \pkg{maptools} packages are available.
+#'   Note that the \href{https://CRAN.R-project.org/package=alphahull}{alphahull} package
+#'   is released under a restrictive non-free software license.
 #' @param width 'numeric' number.
 #'   Buffer distance from geometry of convex hull.
 #' @param ...
@@ -22,7 +25,8 @@
 #'   On the shape of a set of points in the plane:
 #'   IEEE Transactions on Information Theory, v. 29, no. 4, p. 551--559.
 #'
-#' @seealso \code{\link[grDevices]{chull}}, \code{\link[alphahull]{ashape}}
+#' @seealso \code{\link[grDevices]{chull}}, \code{\link[alphahull]{ashape}},
+#'   \code{\link[maptools]{checkPolygonsHoles}}
 #'
 #' @keywords utilities
 #'
@@ -30,21 +34,23 @@
 #'
 #' @examples
 #' set.seed(123)
-#' pts <- sp::SpatialPoints(cbind(x = stats::runif(50), y = stats::runif(50)))
-#' sp::plot(GetRegionOfInterest(pts, width = 0.05), border = "red", lty = 2)
+#'
+#' n <- 50
+#' pts <- sp::SpatialPoints(cbind(x = stats::runif(n), y = stats::runif(n)))
+#' sp::plot(GetRegionOfInterest(pts, width = 0.05), border = "blue", lty = 2)
 #' sp::plot(GetRegionOfInterest(pts), border = "red", add = TRUE)
-#' sp::plot(GetRegionOfInterest(pts, alpha = 0.5), border = "blue", add = TRUE)
+#' sp::plot(GetRegionOfInterest(pts, width = -0.05), lty = 2, add = TRUE)
 #' sp::plot(pts, add = TRUE)
 #'
+#' \dontrun{
 #' n <- 300
-#' set.seed(321)
 #' theta <- stats::runif(n, 0, 2 * pi)
 #' r <- sqrt(stats::runif(n, 0.25^2, 0.50^2))
-#' x <- cbind(0.5 + r * cos(theta), 0.5 + r * sin(theta))
-#' pts <- sp::SpatialPoints(x)
+#' pts <- sp::SpatialPoints(cbind(0.5 + r * cos(theta), 0.5 + r * sin(theta)))
 #' sp::plot(GetRegionOfInterest(pts, alpha = 0.1, width = 0.05), col = "green")
 #' sp::plot(GetRegionOfInterest(pts, alpha = 0.1), col = "yellow", add = TRUE)
 #' sp::plot(pts, add = TRUE)
+#' }
 #'
 
 GetRegionOfInterest <- function(obj, alpha=NULL, width=0, ...) {
@@ -76,13 +82,13 @@ GetRegionOfInterest <- function(obj, alpha=NULL, width=0, ...) {
       stop(sprintf("alpha-shape computation requires the %s package", pkg))
   }
 
-  # code adapted from RPubs post by Barry Rowlingson,
+  # code adapted from RPubs document by Barry Rowlingson,
   # accessed November 15, 2018 at https://rpubs.com/geospacedman/alphasimple
   shp <- alphahull::ashape(coords[, 1:2], alpha=alpha)
   el <- cbind(as.character(shp$edges[, "ind1"]), as.character(shp$edges[, "ind2"]))
   gr <- igraph::graph_from_edgelist(el, directed=FALSE)
   clu <- igraph::components(gr, mode="strong")
-  ply <- lapply(seq_len(clu$no), function(i) {
+  ply <- sp::Polygons(lapply(seq_len(clu$no), function(i) {
     vids <- igraph::groups(clu)[[i]]
     g <- igraph::induced_subgraph(gr, vids)
     if (any(igraph::degree(g) != 2))
@@ -93,9 +99,7 @@ GetRegionOfInterest <- function(obj, alpha=NULL, width=0, ...) {
     idxs <- as.integer(igraph::V(g)[path]$name)
     pts <- shp$x[c(idxs, idxs[1]), ]
     sp::Polygon(pts)
-  })
-  ply <- sp::Polygons(ply, ID=1)
+  }), ID=1)
 
-  ply <- maptools::checkPolygonsHoles(ply)
-  ply
+  maptools::checkPolygonsHoles(ply)
 }