+++ NEWS.md
@@ -1,4 +1,4 @@
-# inlmisc 0.4.4.9000
+# inlmisc 0.4.5
 
 - In `PlotGraph` function, change alignment of `main` title from center to left-hand side of plot.